Minor fixes and improvements
Some of the AVP test cases were missing the "AVP" name in them. This
happened when I was experimenting with having the AVP code in its own
sub-package and forgot to rename them.

There are some readability improvements in the server code, both on
variable names and comments.

The example benchmark tool can now count the total messages and time
taken to run the test when using more than one connection.