Fix for HandlerFunc.ErrorReports
Closes #8.

The only reason for having the ErrorReports method defined on the
HandlerFunc is for the HandlerFunc to satisfy the Handler interface.

Since ServeMux takes either a Handler (via ServeMux.Handle) or a
HandlerFunc (via ServeMux.HandleFunc) to register an endpoint,
internally it always store a Handler.

When ServeMux.HandleFunc is called, the passed HandlerFunc is converted
to a Handler so it can be stored. However, a HandlerFunc can never
report errors, which means its ErrorReports function is only defined to
satisfy the Handler interface and will never be called.