Add credit control RFC and support to the examples
This is an update to the client and server examples that adds support
for the credit control dictionary. As explained in the source code,
these examples only illustrate how the go-diameter package works and do
NOT implement all the commands in the spec. For that you have to
understand both the base and credit control RFCs and implement them
correctly.

The credit control RFC has been added to the repo as well.

This change addresses Issue #9.