Fix constant names to satisfy golint
There's a separate script to generate codes.go now. Not sure if this
should be called automatically using Go generate for now, but it can in
the future.

The focus of this commit is to clean up the file and make it satisfy
golint with no warnings - mostly about Id vs ID.