Auto generating files based on dictionary contents
Currently 3 files are automatically generated from dictionaries that
live in dict/testdata/*.xml:

- commands.go: List of Diameter commands (CE, CC, etc)
- avp/codes.go: List of AVP codes (OriginHost and so on)
- dict/default.go: Default built-in dictionary parser

There's no need to use Go generate for now, so once a dictionary is
updated (although it shoudln't) or a new dictionary is added, we should
run `sh autogen.sh` once. The generated code is gofmt'ed as expected.

Since this commit, the default Parser will include both the Base
dictionary (app=0) and the Credit Control dictionary (app=4).