Add Dictionary to the Conn interface
This adds the Dictionary function to the Conn interface for clients and
servers. If someone is writing a raw client or server they can always
rely on the Conn object to refer back to the dictionary object
associated to the connection. Yet, it's still possible to read and write
messages regardless to having a connection (e.g. for testing).

Both Conn and Message can now return the Dictionary parser associated to
them.