Bugfix for the Conn returned by the client
The client's underlying newConn already creates and initializes a
response object. Instead of returning it (c.writer) we were creating a
new one. This would break the new Context API since the Context saved
in the response/Conn object returned by Dial and DialTLS was not the
same of the Conn object passed to handlers.