Fix for AVPs containing vendor bit
The NewAVP function was expecting that users passing the vendor ID would
also pass the Vbit in the flags. Now there's a check to do that
automatically if the value of vendor is non-zero.

Also, there was a duplicated logic for allocating bytes to serialize the
AVP which has been merged into using avp.Len.

Closes #36.