Release v4.0.0 with go.mod and go.sum
Opting for not creating a subdirectory for version, similar to what was
done in golang.org/x/{net,tools} etc.

These files were created by setting GO111MODULE=on and executing
"go mod init" and "go mod tidy".

There are several extraneous dependencies in go-diameter as of today,
mostly because of examples/s6a_proxy depending on protobuf. In the
future we might consider moving this example out to its own repo to
reduce dependencies.